% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.BOSO}
\alias{coef.BOSO}
\title{Extract coefficients from a BOSO object}
\usage{
\method{coef}{BOSO}(object, beta0 = F, ...)
}
\arguments{
\item{object}{Fitted 'BOSO' or 'BOSO.single' object}

\item{beta0}{Force beta0 to appear (output of 'p+1' features)}

\item{...}{extra arguments for future updates}
}
\value{
A `matrix` object with the corresponding beta values estimated.
}
\description{
This is an equivalent function to the one offered by 
  \code{\link[glmnet]{coef.glmnet}} for extraction of coefficients.
}
