% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.BOSO}
\alias{predict.BOSO}
\title{Predict function for BOSO object.}
\usage{
\method{predict}{BOSO}(object, newx, ...)
}
\arguments{
\item{object}{Fitted 'BOSO' or 'BOSO.single' object}

\item{newx}{Matrix with new data for prediction with BOSO}

\item{...}{extra arguments for future updates}
}
\value{
A `matrix` object with the corresponding beta values estimated.
}
\description{
This is an equivalent function to the one offered by 
  \code{\link[glmnet]{coef.glmnet}} for extraction of coefficients.
}
