\name{SimResultsVignette}
\docType{data}
\alias{sim.results}
\title{sim.results for the vignette}
\description{
  Results from all the algorithms using the simmulated data Simmulated data for the high-5-sized scenario. 
  }
\usage{data("SimResultsVignette")}
\references{
  Hastie, Trevor, Robert Tibshirani, and Ryan J. Tibshirani. "Extended comparisons of best subset selection, forward stepwise selection, and the lasso." arXiv preprint arXiv:1707.08692 (2017).
}
\keyword{datasets}