% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{InternalFunctions}
\alias{InternalFunctions}
\alias{standardize}
\alias{ICscoreCalculation}
\alias{Trace}
\title{BOSO Internal Functions}
\usage{
standardize(
  x,
  y,
  intercept = T,
  normalize = T,
  mx = NULL,
  my = NULL,
  sx = NULL
)

ICscoreCalculation(object, IC, n, p)

Trace(mat)
}
\arguments{
\item{object}{Fitted 'BOSO' or object}

\item{n}{Fitted 'BOSO' object}

\item{p}{Fitted 'BOSO'  object}

\item{metric}{information criteria to be used}

\item{matrix}{a numeric matrix}
}
\value{
Internal outputs
}
\description{
Internal functions used by BOSO in the different steps of the algorithm

Centering and scaling convenience function
based on the function from bestSubset
}
\keyword{internal}
